Simplify ManifestParser
