+++ lib/dash/dash_parser.js
@@ -2232,8 +2232,6 @@ shaka.dash.DashParser.GenerateSegmentIndexFunction;
 shaka.dash.DashParser.StreamInfo;
 
 
-shaka.media.ManifestParser.registerParserByExtension(
-    'mpd', () => new shaka.dash.DashParser());
 shaka.media.ManifestParser.registerParserByMime(
     'application/dash+xml', () => new shaka.dash.DashParser());
 shaka.media.ManifestParser.registerParserByMime(
