+++ lib/hls/hls_parser.js
@@ -3881,8 +3881,6 @@ shaka.hls.HlsParser.PresentationType_ = {
   LIVE: 'LIVE',
 };
 
-shaka.media.ManifestParser.registerParserByExtension(
-    'm3u8', () => new shaka.hls.HlsParser());
 shaka.media.ManifestParser.registerParserByMime(
     'application/x-mpegurl', () => new shaka.hls.HlsParser());
 shaka.media.ManifestParser.registerParserByMime(
