+++ lib/media/manifest_parser.js
@@ -6,9 +6,8 @@
 
 goog.provide('shaka.media.ManifestParser');
 
-goog.require('goog.Uri');
+goog.require('shaka.Deprecate');
 goog.require('shaka.log');
-goog.require('shaka.net.NetworkingEngine');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.Platform');
 
@@ -28,7 +27,9 @@ shaka.media.ManifestParser = class {
    * @export
    */
   static registerParserByExtension(extension, parserFactory) {
-    shaka.media.ManifestParser.parsersByExtension[extension] = parserFactory;
+    shaka.Deprecate.deprecateFeature(5,
+        'ManifestParser',
+        'Please use an ManifestParser with registerParserByMime function.');
   }
 
 
@@ -70,9 +71,6 @@ shaka.media.ManifestParser = class {
       for (const type in ManifestParser.parsersByMime) {
         support[type] = true;
       }
-      for (const type in ManifestParser.parsersByExtension) {
-        support[type] = true;
-      }
     }
 
     // Make sure all well-known types are tested as well, just to show an
@@ -86,14 +84,6 @@ shaka.media.ManifestParser = class {
       // SmoothStreaming
       'application/vnd.ms-sstr+xml',
     ];
-    const testExtensions = {
-      // DASH
-      'mpd': 'application/dash+xml',
-      // HLS
-      'm3u8': 'application/x-mpegurl',
-      // SmoothStreaming
-      'ism': 'application/vnd.ms-sstr+xml',
-    };
 
     for (const type of testMimeTypes) {
       // Only query our parsers for MSE-enabled platforms.  Otherwise, query a
@@ -102,18 +92,6 @@ shaka.media.ManifestParser = class {
         support[type] = !!ManifestParser.parsersByMime[type];
       } else {
         support[type] = shaka.util.Platform.supportsMediaType(type);
-      }
-    }
-
-    for (const extension in testExtensions) {
-      // Only query our parsers for MSE-enabled platforms.  Otherwise, query a
-      // temporary media element for native support for these MIME type for the
-      // extension.
-      if (shaka.util.Platform.supportsMediaSource()) {
-        support[extension] = !!ManifestParser.parsersByExtension[extension];
-      } else {
-        const type = testExtensions[extension];
-        support[extension] = shaka.util.Platform.supportsMediaType(type);
       }
     }
 
@@ -126,12 +104,10 @@ shaka.media.ManifestParser = class {
    * parse the manifest at |uri|.
    *
    * @param {string} uri
-   * @param {!shaka.net.NetworkingEngine} netEngine
-   * @param {shaka.extern.RetryParameters} retryParams
    * @param {?string} mimeType
    * @return {shaka.extern.ManifestParser.Factory}
    */
-  static getFactory(uri, netEngine, retryParams, mimeType) {
+  static getFactory(uri, mimeType) {
     const ManifestParser = shaka.media.ManifestParser;
 
     // Try using the MIME type we were given.
@@ -145,19 +121,6 @@ shaka.media.ManifestParser = class {
           'Could not determine manifest type using MIME type ', mimeType);
     }
 
-    const extension = ManifestParser.getExtension(uri);
-    if (extension) {
-      const factory = ManifestParser.parsersByExtension[extension];
-      if (factory) {
-        return factory;
-      }
-
-      shaka.log.warning(
-          'Could not determine manifest type for extension ', extension);
-    } else {
-      shaka.log.warning('Could not find extension for ', uri);
-    }
-
     throw new shaka.util.Error(
         shaka.util.Error.Severity.CRITICAL,
         shaka.util.Error.Category.MANIFEST,
@@ -167,46 +130,21 @@ shaka.media.ManifestParser = class {
 
 
   /**
-   * @param {string} uri
-   * @return {string}
-   */
-  static getExtension(uri) {
-    const uriObj = new goog.Uri(uri);
-    const uriPieces = uriObj.getPath().split('/');
-    const uriFilename = uriPieces.pop();
-    const filenamePieces = uriFilename.split('.');
-
-    // Only one piece means there is no extension.
-    if (filenamePieces.length == 1) {
-      return '';
-    }
-
-    return filenamePieces.pop().toLowerCase();
-  }
-
-
-  /**
-   * Determines whether or not this URI and MIME type are supported by our own
+   * Determines whether or not the MIME type is supported by our own
    * manifest parsers on this platform.  This takes into account whether or not
    * MediaSource is available, as well as which parsers are registered to the
    * system.
    *
-   * @param {string} uri
    * @param {string} mimeType
    * @return {boolean}
    */
-  static isSupported(uri, mimeType) {
+  static isSupported(mimeType) {
     // Without MediaSource, our own parsers are useless.
     if (!shaka.util.Platform.supportsMediaSource()) {
       return false;
     }
 
     if (mimeType in shaka.media.ManifestParser.parsersByMime) {
-      return true;
-    }
-
-    const extension = shaka.media.ManifestParser.getExtension(uri);
-    if (extension in shaka.media.ManifestParser.parsersByExtension) {
       return true;
     }
 
@@ -255,13 +193,5 @@ shaka.media.ManifestParser.AccessibilityPurpose = {
  * @type {!Object.<string, shaka.extern.ManifestParser.Factory>}
  */
 shaka.media.ManifestParser.parsersByMime = {};
-
-
-/**
- * Contains the parser factory functions indexed by file extension.
- *
- * @type {!Object.<string, shaka.extern.ManifestParser.Factory>}
- */
-shaka.media.ManifestParser.parsersByExtension = {};
