+++ lib/mss/mss_parser.js
@@ -1077,7 +1077,5 @@ shaka.mss.MssParser.Context;
  */
 shaka.mss.MssParser.TimeRange;
 
-shaka.media.ManifestParser.registerParserByExtension(
-    'ism', () => new shaka.mss.MssParser());
 shaka.media.ManifestParser.registerParserByMime(
     'application/vnd.ms-sstr+xml', () => new shaka.mss.MssParser());
