+++ lib/offline/storage.js
@@ -279,8 +279,6 @@ shaka.offline.Storage = class {
 
       const factory = shaka.media.ManifestParser.getFactory(
           uri,
-          this.networkingEngine_,
-          config.manifest.retryParameters,
           mimeType || null);
 
       return factory();
