+++ lib/player.js
@@ -7,8 +7,9 @@
 goog.provide('shaka.Player');
 
 goog.require('goog.asserts');
-goog.require('shaka.Deprecate');
+goog.require('goog.Uri');
 goog.require('shaka.config.AutoShowText');
+goog.require('shaka.Deprecate');
 goog.require('shaka.log');
 goog.require('shaka.media.AdaptationSetCriteria');
 goog.require('shaka.media.BufferingObserver');
@@ -1278,7 +1279,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
     // If we don't have a MIME type, try to guess based on the file extension.
     // TODO: Too generic to belong to ManifestParser now.  Refactor.
     // Try using the uri extension.
-    const extension = shaka.media.ManifestParser.getExtension(this.assetUri_);
+    const extension = this.getExtension_(this.assetUri_);
     if (shaka.Player.EXTENSIONS_TO_MIME_TYPES_[extension]) {
       return shaka.Player.EXTENSIONS_TO_MIME_TYPES_[extension];
     }
@@ -1299,6 +1300,26 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       mimeType = 'application/vnd.apple.mpegurl';
     }
     return mimeType;
+  }
+
+
+  /**
+   * @param {string} uri
+   * @return {string}
+   * @private
+   */
+  getExtension_(uri) {
+    const uriObj = new goog.Uri(uri);
+    const uriPieces = uriObj.getPath().split('/');
+    const uriFilename = uriPieces.pop();
+    const filenamePieces = uriFilename.split('.');
+
+    // Only one piece means there is no extension.
+    if (filenamePieces.length == 1) {
+      return '';
+    }
+
+    return filenamePieces.pop().toLowerCase();
   }
 
   /**
@@ -1334,7 +1355,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       }
 
       const canPlayMediaSource =
-          shaka.media.ManifestParser.isSupported(assetUri, mimeType);
+          shaka.media.ManifestParser.isSupported(mimeType);
 
       // If MediaSource isn't an option, the native option is our only chance.
       if (!canPlayMediaSource) {
@@ -1457,9 +1478,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       // Create the parser that we will use to parse the manifest.
       this.parserFactory_ = shaka.media.ManifestParser.getFactory(
           this.assetUri_,
-          networkingEngine,
-          this.config_.manifest.retryParameters,
-          mimeType || null);
+          mimeType);
       goog.asserts.assert(this.parserFactory_, 'Must have manifest parser');
       this.parser_ = this.parserFactory_();
 
@@ -4669,7 +4688,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
    */
   async getTextMimetype_(uri) {
     // Try using the uri extension.
-    const extension = shaka.media.ManifestParser.getExtension(uri);
+    const extension = this.getExtension_(uri);
     let mimeType = shaka.Player.TEXT_EXTENSIONS_TO_MIME_TYPES_[extension];
 
     if (mimeType) {
